back-end/requirements.txt
flask
pillow
tifffile
flask-cors
Flask-SQLAlchemy
python-dotenv